Check earlier if application throws error
If an application produces an error it is not eligible to shadow other
applications either by name or by path.

The current approach creates the application twice, first to check if
the application is healthy and, if the application is eligible for
registration, it is finally registered.

This approach is needed because it is not possible to create a bus for
the application, register the bus in a servlet and later unregister the
servlet. When the servlet is destroyed the bus is also destroyed.

Also it is not trivial for me to perform the shadowing checks and, if the
application errors, redo the checks and keep the program reusable and
untangled.