Proper extension ordering following the spec
We need to use CXF models directly in order to be able to sort by
ServiceReference in case of a tie.

Basically in this I am copying the same registration as the Feature did
but using our own ProviderInfo subclass containing the
ServiceReference. This way the provider ProviderInfoComparator can solve
the ties.