Allow whiteboard to target ServletContextHelper
If the whiteboard instance configuration target one specific HTTP
whiteboard ServletContextHelper, all applications that don't specify one
whiteboard themselves will use the one in the whiteboard configuration.

If no ServletContextHelper is specified one would be created for the
incoming application as usual.

Specifying a ServletContextHelper, both in the whiteboard configuration
or in the incoming application service properties, breaks application
isolation.