[ARIES-1929] Special error handling in the first rewire
If we get a wiring error in the first (probably bulk) rewire we need to
discard one by one the possible services that are causing it. Since CXF
does not inform about which service is causing the error we need to try
incrementally. Once we have ruled out the failing services we mark them
as such in the runtime and in the extension registry, which might cause
some more extensions or services to be unregistered from the
application (even the very application).

During this process we disable the registrator so we don't get
re-entrant rewires and, when finalized, we try to wire the application
again. If the application still fails we mark it as failing.