[ARIES-1928] Track extensions per application
Before there was a global registry for extensions. The presence or
absence of an extension does not mean anything until it is initialized
inside an application. An application depending on extensions should not
be deployed until its dependencies have been checked to work.

Also, instead of registering the CxfJaxrsServiceRegistrator, we now
register the application in a internal registry, which simplifies the
extension attaching.