Async invocation performance improvements
Despite having an asynchronous option, the default CXF client is purely synchronous. Furthermore, the more advanced CXF clients are still synchronous if a custom SSLContext is used. Finally, even when asynchronous, all of the filters, interceptors, readers and writers are run on the calling thread. Performance is better with a simpler implementation which makes use of the PromiseFactory#submit() method and a SyncInvoker.

Signed-off-by: Tim Ward <timothyjward@apache.org>